// src/app/models/pokemon-type.ts
export enum PokemonType {
    NORMAL = 'normal',
    FIGHTING = 'fighting',
    FLYING = 'flying',
    POISON = 'poison',
    GROUND = 'ground',
    ROCK = 'rock',
    BUG = 'bug',
    GHOST = 'ghost',
    STEEL = 'steel',
    FIRE = 'fire',
    WATER = 'water',
    GRASS = 'grass',
    ELECTRIC = 'electric',
    PSYCHIC = 'psychic',
    ICE = 'ice',
    DRAGON = 'dragon',
    DARK = 'dark',
    FAIRY = 'fairy',
}

export const PokemonTypeColors: Record<PokemonType, string> = {
    [PokemonType.NORMAL]: '#808080',
    [PokemonType.FIGHTING]: '#C03028',
    [PokemonType.FLYING]: '#A890F0',
    [PokemonType.POISON]: '#A040A0',
    [PokemonType.GROUND]: '#E0C068',
    [PokemonType.ROCK]: '#B8A038',
    [PokemonType.BUG]: '#A8B820',
    [PokemonType.GHOST]: '#705898',
    [PokemonType.STEEL]: '#B8B8D0',
    [PokemonType.FIRE]: '#F08030',
    [PokemonType.WATER]: '#6890F0',
    [PokemonType.GRASS]: '#78C850',
    [PokemonType.ELECTRIC]: '#F8D030',
    [PokemonType.PSYCHIC]: '#F85888',
    [PokemonType.ICE]: '#98D8D8',
    [PokemonType.DRAGON]: '#7038F8',
    [PokemonType.DARK]: '#705848',
    [PokemonType.FAIRY]: '#EE99AC',
};